requirements.txt
tensorflow==2.12.0
numpy==1.23.5
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
jinja2==3.1.2 